requirements.txt
requests>=2.27.1
numpy>=1.23.0
pandas>=0.23.4
gdown>=3.10.1
tqdm>=4.30.0
Pillow>=5.2.0
opencv-python>=4.5.5.64
tensorflow>=2.10.0
keras>=2.2.0
Flask>=1.1.2
flask_cors>=4.0.1
mtcnn>=0.1.0
retina-face>=0.0.1
fire>=0.4.0
gunicorn>=20.1.0
pinecone>=3.0.0
fastapi[all]==0.110.0
uvicorn==0.29.0
firebase-admin==6.3.0
deepface==0.0.93
python-dotenv>=1.0.0
firebase-functions==0.4.2
tf-keras>=2.10.0
transformers>=4.36.0
torch>=2.0.0
